Replace 'LedgerState' with simply 'UTxO'
In a Hydra Head, we do really only intend to keep UTxO for a ledger
state. Anything more, would not be able to be "re-created" onto the main
chain. So having only a 'UTxO' type for a 'Tx' suggests that not more
should be kept around by a 'Ledger' implementation.