Change API types to not use manual JSON instances
This should result in less code maintenance and golden tests + the
api.yaml should avoid accidental regressions.

Furthermore we deliberately went for full on `Generic` instances as
there were many others like this in the codebase and we should keep it
consistent (move everything at once to some different 'tag' if we want).