Throw PostTxError when tx submission fails #279
This code does some arguably ugly things to ensure the failures in the
tx submission thread are thrown into the postTx callback's thread in
order to be caught at the Node level and not crash. The TQueue for
submission should probably be simplified to a TMVar or something as
evertyhing is synchronous now.