Update without-docker.md
The current version does not run correctly as is. The `/devnet/ipc/` directory is not create. A fix is to replace the node.socket location to the default `/devnet/db`location. Another fix is to create the `/devnet/ipc` directory and use the previous command (this is easier and more compatible with the other documentation.