Do not adjust fanout lower bound
We added a grace period of 100 slots meaning the transaction is very
likely to be rejected when we post it, esp. in a test setting. As ETE
test is also failing because we don't wait enough, this should make it
possible to have the ETE tests pass but we should have a better story
for resubmitting transactions.