Query at point if requested
The queryXXX functions of the includec CardanoClient do use the local
state query protocol, which can acquire some specific point on the chain
when passed `Just ChainPoint`, so we do configure these functions to
take a `QueryPoint`.

We do use this to query at the point to which the chain rolled back in
TinyWallet workflow.