Start with a Nothing instead of AtGenesis chain point
This makes it the same case as no state loaded from persistence and is
more honest in regards to the semantics of 'ChainStateAt'. In the
beginning, before we observe anything, we have actually no information
on the chain point and when the starting state 'Idle' was actually
recorded -> thus 'Nothing'.