Refactor head state mutation
We introduce replaceSomething to ease the mutation of a head state.
This functions are exposed by Mutation.hs as it seems to be the better
place to me right now but maybe we can do better?

Also, we replace some functions which did that and failed when the state
was not an expected one. Now we just return the state unchanged if the
field we want to change does not exist for this state.

It may be a good idea... it also may introduce difficulties in
diagnostic test errors. That is not clear at this point. If ist becomes
a problem, we can sill introduce the errors back easily in the code.