Make script failure in wallet a dedicated error
Ideally though we would have used the cardano-api ScriptExecutionError,
but conversion from ledger types to this is actually not exported and we
want to avoid vendoring this conversion code just for identifying this
case. It's a long standing TODO to switch to the cardano-api types and
balancing code in the wallet anyways.