Build modules in parallel
Quoting @mpj on this:

Cabal does component level parallelism, but not module level
parallelism.

Seems like putting -j is not a good idea as cabal would end up doing
num_parallel_components*num_parallel_modules modules at once, which can
be a lot. And it turns out it was not faster than using a number <
actual available cpus.

We put -j8 as it showed the best performance.