Remove TxSeen and TxExpired
The server outputs TxValid and TxInvalid shall be used instead.

Rationale:

Before we were checking against the confirmed ledger on NewTx and yield
a TxValid if ok, but that is a useless check as it needs to be
applicable to the seen ledger to continue in the protocol. Instead we
now only yield TxValid if it was applicable against the seen ledger.

TxExpired was not giving any error information why it was expired.
Similarly, TxInvalid had no real value as it was in response of checking
against the confirmed ledger, while the seen ledger might already have
advanced and the transaction is actually not applicable. Instead we now
return a TxInvalid in this case with the last validationError that we
saw.