FIX performance issue
Forgot to purge the allTxs list when transaction is included in a
snapshot.

Added a test to ensure that we check that both, when we're supposed
to send a ReqSn and when we're not.

This is not pretty as we duplicate one test so that we can check
for every branch but it's safer now that we're explicit about this
and we can, hopefully, improve that when we migrate to event-sourced
approach.