Move cleanup of allTxs to ReqSn handling
Transactions included in a snapshot can be removed from allTxs.

Before, it was done when the snapshot was fully validated but, in
fact, as soon as we receiver a request for snapshot, we can not
unsee this snapshot so, anyway, we're commited to, either, integrate
these transactions in this snapshot and validate it or kill the
head liveness.