Only keep a note about removing transactions from allTxs
We could not express a behavior test which would motivate this
change and it's not covered by the specification. However, as we can
formulate a situation where we would not want them removed, we should
keep them for now.

If performance because of a growing allTxs becomes an issue, we should
put an upper bound on how many invalid txs we keep in allTxs.