Handle inline script datums
Alter the existing `datum` field of `ScriptInfo` to be optional.
This should be used for datum hashes only from now on and in case
this field is not populated we assume the user is using inline datums.

Move the datum construction outside of `createOutputAtAddress` so that
we have better control of the used datums.