Only resend messages upon receiving a Ping
This is meant to ensure we only try to resend messages whenever the peer is quiescent,
which was the original intent of using Pings in the first place in order to avoid
resending messages too often. The assumption is that disconnections and messages
drop should be few and far between in normal operations and it's therefore fine to
rely on the Ping's roundtrip time to check for peers state.