Fix concurrent update of vector clocks
We removed the use of TVar but callback is not thread safe and
therefore we need to ensure updates to the acks are atomic across
broadcast and callback calls, which is not possible when writing to
disk in both calls.

Broadcast is "guaranteed" to be called from a single thread and
therefore it's the only place where we can safely write to disk.