Add some FIXME notes to observeCommitTx before refactoring it
We plan to use the observeCommitTx also from the chain observer. To make
that happen we need to remove stateful information like the known
initials of a head.

However, those are (in a contrived way) currently our only way of
ensuring that this transaction is a proper Hydra Head transaction.

The FIXME describes alternatives how to achieve this via keeping the
seed in the datum or providing the full resolved inputs of a
transaction (including the committed UTxO).