Delay TxSubmission failure and ignore errors in logic
By not immediately calling back upon having a transaction rejected, the
chain sync client gets to process a block which invalidated our
submitted transaction. In turn, this allows the head logic to safely
ignore some of these errors, e.g. when it does not matter who opens a
head with a collectCom transaction