ServerOutput contains full transactions
This also updates TxValid to only include the transaction id, such that
transactions are only submitted once per client still. Hence, this will
not change the overall bandwidth requirement on websocket clients, but
will make their implementation significantly easier.

Before, if clients wanted to act on transactions this was a lot easier
to do upon seing `TxValid`. However, this was only confirming local
ledger application and not consensus / enforcability of this transaction
onto the L1.

The new API suggests to do the right thing by making it straight-forward
to act upon seeing a transaction in a `SnapshotConfirmed`.