Drop redundant test and add known issues about static topology
The test was asserting that the network layer fails to start with wrong
configuration. The etcd-based network is not strictly failing because of
that, but other issue with the cluster can happen which we recorded in
the known issues. Furthermore, there is also the peer
misconfiguration (via the `state` file) which prevents changing the
`--peer` configuration after successful initiation of a head.