Fix incremental commit/decommit state events
These have not been carrying and updating the 'chainState' in our
'HeadState' aggregate. This would result in the chain layer not
restarting from these last known chain points and we saw the hydra-node
re-observing deposits, incrementas and (likely) decrements.

Thi also re-order cases in aggregate/aggregateHistory to see better that
we ought to use chainState in some of them.