Hotfix Checkpoint handling of API server projections (#2152)
The projections in the API server were not correctly handing
`Checkpoint` events (introduced with event log rotation).

I tried to create a regression test, but we don't have a `Gen
[StateChanged tx]` generators that honors the constraints required for
`aggregate` so I could not get hold of an expected list of
`StateChanged` events with a consistent `Checkpoint` to do property
testing.

The whole `Projection` business is a bit contrived since we have the
full `StateChanged` event stream available (the API server is now an
`EventSink`). As we are also not needing to track changes onto the
projected values (resource-specific websockets / subscriptions would
require that), we should consider dropping the whole mechanism and just
stick with a `getHeadState :: STM (HeadState tx)` + getter functions (or
lenses) to acquire the relevant things in the HTTP and WS API.

This PR though is a hotfix and I kept the diff to a minimum (especially
as its not covered by tests!)

---

* [x] CHANGELOG updated
* [x] Documentation update not needed
* [x] Haddocks updated
* [ ] No new TODOs introduced
  - An XXX note on the same as above