use FIFO as deposit selection order for ReqSn (#2263)
<!-- Describe your change here -->

🔑 **Motivation** 
> Follow up on #2217

When multiple deposits become active at the same time, the node
currently selects the next deposit for inclusion in `ReqSn` based on the
`Ord` instance of `TxId`.

This behavior is not intuitive from a user’s perspective and may lead to
deposits being picked in an order unrelated to their actual arrival
on-chain.

To make the behavior predictable and fair, the selection should follow
FIFO semantics, based on the `created` field of each deposit.

🔄 **Changes**

* Pending active deposits are now picked in arrival order (`created`)
when building `ReqSn` in the `Open` state.

📝 **Notes**

* This change ensures that users see deposits included in the order they
were observed on-chain, improving transparency and reducing surprises.

---

<!-- Consider each and tick it off one way or the other -->
* [x] CHANGELOG updated or not needed
* [X] Documentation updated or not needed
* [X] Haddocks updated or not needed
* [X] No new TODOs introduced or explained herafter